introduced isEqual method. I had it before, but removed it too eagerly in favor of simplicity. Alas, we need two kinds of equality. The reason is this:
{1} \ {1} == {}

the left-hand side produces a set[int], while the right-hand side produces a set[void]. The sets are equal and should be, but their representations as objects are different because their types are different. To facilitate implementations that use this difference (i.e. remove common sub-expressions, or sharing), one must have access to both kinds of equality. One equals is only used in implementations of IValue (equals()), the other is used for computing with values (isEqual()).