Minor changes to enable more liberal treatment of tuple types
It seems that the PDB types could, in principle, handle polymorphic
tuples, e.g. parameter types of the form &T <: tuple[str name] where the
order of fields is immaterial. However, this currently givess various
run-time errors and I tarted to explore the options and commit on the
main branch because I forgot to branch beforehand. These changes do NOT
modify the standard behaviour. All PDB tests run, except for an
unrelated one.

The changes are:
- Added getFieldNames in Type.
- Added (but commented out) an experimental version of isSubTypeOf in
TupleType.
- Added a version of equivalent in TupleType that is called in
overloading resolution