Major change in build: switching from maven/tycho to maven-bundle-plugin.
This change guarantees to use the same (e.g. maven) build configuration on command line and in different ideas (i.e. Eclipse, IntelliJ IDEA).
One of the implications is, that dependencies are now resolved via maven and not anymore over MANIFEST.MF referenced p2 / OSGi component repositories.

It still has to be monitored if this build setting setup behaves (experimental).