PDB Binary Writer got smarter when writing with compression = false
It now at least checks if the identity is the same, such that those
references are kept after re-reading the file.
Stuff like parse tree and certain m3 models would blow up in size.

Sadly there is some strange code in both the reader and the writer
to never cache the values with tags, so to keep compatibility,
so we can only apply the this improvement for values without tags.