Moved the FunctionType definition and implementation from the rascal project to the vallang project
The reason is that TupleType and FunctionType shared a lot in common,
but not everything. This made FunctionType in the rascal project brittle
and the abstraction was not safe. Now FunctionType is still implemented
using TupleType, but some features of tuples are judiciously turned-off
for parameter types and keyword parameter types. Namely the reduction
to void if one of the parameter fields is void was not correct
for parameter lists, while it is essential for real tuple types.

FunctionType currently does not come with the corresponding IFunction
extends IValue interface. This is for the near future, when we've
figured out how to (de)serialize implementations of IFunction in a
generic and readable fashion (as all vallang values have).