Solve perf issue serializing large collections
We have a stackless structured visitor that doesn't require large call stacks to deal with deeply nested ASTs.

While this worked great for deep ASTs, it doesn't work great for big flat collections (aka wide).

Big collections would require the same amount of memory, just to prepare the stack with all the entries.
Now we have iterating entries on the stack, that can be "returned to" multiple times.